+++ src/pypolca/utils/fix_consensus_from_vcf.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 
+import re
 from pathlib import Path
 from typing import Dict
 
@@ -24,18 +25,30 @@ def fix_consensus_from_vcf(ref_contigs: Path, vcf: Path, out_fasta: Path) -> Non
     """
 
     rseq = {}
+    ctg, seq = "", ""
+    rseq = {}
 
-    # Read reference sequences into memory
-    with open(ref_contigs) as fasta_file:
-        for record in SeqIO.parse(fasta_file, "fasta"):
-            rseq[record.id] = str(record.seq)
+    with open(ref_contigs, "r") as file:
+        for line in file:
+            line = line.strip()
+            if line.startswith(">"):
+                if seq:
+                    rseq[ctg] = seq
+                ctg = line[1:]
+                seq = ""
+            else:
+                seq += line
+
+    if seq:
+        rseq[ctg] = seq
 
     # Initialize variables
     ctg = ""
     fixes = []
     originals = []
     offsets = []
-    fixed_sequences = {}
+
+    total_count = 0
 
     # Read and process VCF file
     with open(vcf) as vcf_file:
@@ -43,14 +56,17 @@ def fix_consensus_from_vcf(ref_contigs: Path, vcf: Path, out_fasta: Path) -> Non
             line = line.strip()
             if line.startswith("#"):
                 continue
-            fields = line.split()
-            if "," in fields[4] or fields[0] not in rseq:
+
+            f = line.split()
+            if "," in f[4] or f[0] not in rseq:
                 continue
-            if fields[0] != ctg:
+
+            # if the contig is new it will instantaite a new contig - is reading line by line so needed
+            if f[0] != ctg:
                 if fixes:
                     if ctg not in rseq:
                         raise Exception(
-                            "Sequence {} not found in the input fasta file".format(ctg)
+                            f"sequence {ctg} not found in the input fasta file"
                         )
                     oldseq = rseq[ctg]
 
@@ -63,68 +79,75 @@ def fix_consensus_from_vcf(ref_contigs: Path, vcf: Path, out_fasta: Path) -> Non
                             offsets[i] - 1 : offsets[i] - 1 + len(originals[i])
                         ]
                         if (
-                            set(original_seq).intersection("acgtnACGTN")
-                            and not originals[i].upper() == original_seq.upper()
+                            any(c in "acgtnACGTN" for c in original_seq)
+                            and not original_seq.upper() == originals[i].upper()
                         ):
                             logger.warning(
-                                "WARNING! Sequence does not match the original {} {} {} {}".format(
-                                    ctg, original_seq, originals[i], offsets[i]
-                                )
+                                "WARNING! Sequence does not match the original:",
+                                ctg,
+                                original_seq,
+                                originals[i],
+                                offsets[i],
                             )
                         else:
-                            # Then, substitute
+                            # Then substitute
                             oldseq = (
                                 oldseq[: offsets[i] - 1]
                                 + fixes[i]
                                 + oldseq[offsets[i] - 1 + len(originals[i]) :]
                             )
 
-                    fixed_sequences[ctg] = oldseq
+                    rseq[ctg] = oldseq
 
                 fixes = []
                 originals = []
                 offsets = []
-                ctg = fields[0]
+                ctg = f[0]
 
-            ff = fields[9].split(":")
-            if int(ff[5]) > 1 and int(ff[5]) >= 2 * int(ff[3]):
-                fixes.append(fields[4])
-                originals.append(fields[3])
-                offsets.append(int(fields[1]))
+            # append if meets criteria for POLCA
+            ff = f[9].split(":")
+            if int(ff[5]) > 1:
+                if int(ff[5]) >= 2 * int(ff[3]):
+                    fixes.append(f[4])
+                    originals.append(f[3])
+                    offsets.append(int(f[1]))
+                    total_count += 1
 
-    if len(fixes) > 0:
+    # actually fix the report now
+    # if fixes as previously means that this step wouldn't continue if the last contig had no changes
+    # therefore, use a total count variable to achieve this.
+    if total_count > 0:
         logger.info(f"POLCA has found variants. Fixing")
-        # Proceed with fixing for the last contig
-        if ctg not in rseq:
-            raise Exception("Sequence {} not found in the input fasta file".format(ctg))
+        # Proceed with fixing
         oldseq = rseq[ctg]
-
         for i in range(
             len(fixes) - 1, -1, -1
         ):  # Going in reverse order to avoid shifting sequence due to indels
-            # First, we check if the sequence at the given offset matches the original variant
+            if ctg not in rseq:
+                raise Exception(f"sequence {ctg} not found in the input fasta file")
             original_seq = oldseq[offsets[i] - 1 : offsets[i] - 1 + len(originals[i])]
             if (
-                set(original_seq).intersection("acgtnACGTN")
-                and not originals[i].upper() == original_seq.upper()
+                any(c in "acgtnACGTN" for c in original_seq)
+                and not original_seq.upper() == originals[i].upper()
             ):
                 logger.warning(
-                    "WARNING! Sequence does not match the original {} {} {} {}".format(
-                        ctg, original_seq, originals[i], offsets[i]
-                    )
+                    "WARNING! Sequence does not match the original:",
+                    ctg,
+                    original_seq,
+                    originals[i],
+                    offsets[i],
                 )
             else:
-                # Then, substitute
                 oldseq = (
                     oldseq[: offsets[i] - 1]
                     + fixes[i]
                     + oldseq[offsets[i] - 1 + len(originals[i]) :]
                 )
 
-        fixed_sequences[ctg] = oldseq
+        rseq[ctg] = oldseq
 
         records = []
-        for contig, sequence in fixed_sequences.items():
+        for contig, sequence in rseq.items():
             record = SeqIO.SeqRecord(seq=Seq(sequence), id=contig, description="")
             records.append(record)